refactor(mm-next): fix filename typo
