+++ packages/mirror-media-next/components/login/main-form-registration.js
@@ -17,7 +17,7 @@ import { createUserWithEmailAndPassword } from 'firebase/auth'
 import { FirebaseError } from 'firebase/app'
 import EmailInput from './email-input'
 import RegistrationPasswordInput from './registration-password-input'
-import CenteredHint from '../shared/cemtered-hint'
+import CenteredHint from '../shared/centered-hint'
 import { InputState } from '../../constants/component'
 import PrimaryButton from '../shared/buttons/primary-button'
 import DefaultButton from '../shared/buttons/default-button'
